requirements.txt
# BE
Flask==2.2.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
Pillow==9.3.0
Flask-Cors==3.0.10
PyJWT==2.6.0

# AI
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.1
requests==2.28.1
torch==1.9.0
torchvision==0.10.0